// platformx/libs/utilities/src/lib/hooks/usePrelemImpression/usePrelemImpression.types.tsx
export interface secondaryArgsObj {
  prelemImpressionSchema?: string;
  clickImpressionSchema?: string;
}
export interface Analytics {
  pageId?: number;
  prelemId?: number;
  pageTitle?: string;
  prelemTitle?: string;
  pageDesc?: string;
  pageTags?: string;
  prelemTags?: string;
  prelemPosition?: number;
  isAnalyticsEnabled: boolean;
  isAuthoring: boolean;
  isSeoEnabled: boolean;
}

export interface Analytics {
  pageId?: number;
  prelemId?: number;
  pageTitle?: string;
  prelemTitle?: string;
  pageDesc?: string;
  pageTags?: string;
  prelemTags?: string;
  prelemPosition?: number;
  isAnalyticsEnabled: boolean;
  isAuthoring: boolean;
  isSeoEnabled: boolean;
}
export interface ButtonObjInfo {
  Button_Action?: string;
  Button_Content?: string;
  Button_Name?: string;
  Button_RedirectURL?: string;
  Button_RestEndPoint?: string;
  Button_Type?: string;
  Button_Value?: string;
}
export interface CardDataObj {
  prelemSlotNumber?: string | number;
  contentType?: string;
  contentTitle?: string;
  contentUrl?: string;
  isRegistered?: string;
  age?: string;
  gender?: string;
}

export interface secondaryArgsObj {
  prelemImpressionSchema?: string;
  clickImpressionSchema?: string;
}